+++ generator/src/main/java/com/scottlogic/deg/generator/generation/rows/GeneratedObjectMerger.java
@@ -1,6 +1,5 @@
 package com.scottlogic.deg.generator.generation.rows;
 
-import com.scottlogic.deg.generator.Value;
 import com.scottlogic.deg.generator.Field;
 import com.scottlogic.deg.generator.FlatMappingSpliterator;
