+++ generator/src/main/java/com/scottlogic/deg/generator/generation/rows/RowSourceFactory.java
@@ -1,7 +1,6 @@
 package com.scottlogic.deg.generator.generation.rows;
 
 import com.google.inject.Inject;
-import com.scottlogic.deg.generator.Value;
 import com.scottlogic.deg.generator.Field;
 import com.scottlogic.deg.generator.fieldspecs.FieldSpec;
 import com.scottlogic.deg.generator.fieldspecs.RowSpec;
