+++ generator/src/main/java/com/scottlogic/deg/generator/walker/reductive/ReductiveTreePruner.java
@@ -1,7 +1,7 @@
 package com.scottlogic.deg.generator.walker.reductive;
 
 import com.google.inject.Inject;
-import com.scottlogic.deg.generator.Value;
+import com.scottlogic.deg.generator.generation.rows.Value;
 import com.scottlogic.deg.generator.Field;
 import com.scottlogic.deg.generator.constraints.atomic.AtomicConstraint;
 import com.scottlogic.deg.generator.constraints.atomic.AtomicConstraintsHelper;
