+++ generator/src/test/java/com/scottlogic/deg/generator/walker/reductive/ReductiveFieldSpecBuilderTests.java
@@ -1,6 +1,6 @@
 package com.scottlogic.deg.generator.walker.reductive;
 
-import com.scottlogic.deg.generator.Value;
+import com.scottlogic.deg.generator.generation.rows.Value;
 import com.scottlogic.deg.generator.Field;
 import com.scottlogic.deg.generator.constraints.atomic.IsNullConstraint;
 import com.scottlogic.deg.generator.decisiontree.TreeConstraintNode;
