+++ generator/src/test/java/com/scottlogic/deg/generator/walker/reductive/ReductiveStateTests.java
@@ -1,6 +1,6 @@
 package com.scottlogic.deg.generator.walker.reductive;
 
-import com.scottlogic.deg.generator.Value;
+import com.scottlogic.deg.generator.generation.rows.Value;
 import com.scottlogic.deg.generator.Field;
 import com.scottlogic.deg.generator.ProfileFields;
 import org.junit.jupiter.api.Test;
